Rename files, correcting typo
